Switch to opensource modemd [2/2]
This is to remedy (not prevent or fix) the modem failure that occurs when having
two sims at dsds. It allows modemd to peacefully reset the modem when it fails.

This removes the modemd from the binary blobs.

This commit is split to two. Check the other on device_samsung_kanas.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>