Remove modemd from vendor blobs.
We'll use the opensource modemd for one reason, it does not make the phone
trigger a kernel panic or restart. I'm also not 100% sure whether our vendor blobs
would never do a reboot in all possible scenarios since I do not have the code.

This commit is split to two. Check the other on device_samsung_kanas.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>