kanas: Prepare script for migration to /vendor/ on Android O
We don't have /vendor/ partition but does seem to be the norm for Oreo and above versions.

I have to redo this branch since I forgot to push the most recent (June) commits before
my hard drive failed (on December) and took all my files repos and builds with it.
* cries *

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>