kanas: Update file paths to point to /vendor/ partition
This was easily done using a hex editor like bless.

We change all binaries and libraries to point to their dependencies in /vendor/.
Fortunately /vendor/ and /system/ are of the same bytes, so we can just overwrite most /system/ references to /vendor/.
We do not write all /system/ to /vendor/ since some files like /system/bin/linker should always be in the /system/.

If any blob became bigger or smaller even by a single byte, I may have made some misteak.

Also, /system/csc paths are left as is.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>