kanas: gps: rename gps.default.so to gps.sc8830.so for the wrapper to work
I wrote a wrapper from a reference to make the gps work again but it needs to take the
name of the gps.default.so for it to do its wonders.

Let's rename this HAL to some inconspicuous file name such as gps.sc8830.so so that it won't
conflict.
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>

Change-Id: Ic0d1e46aebe92346b7974bed01c06772a96cfde6