gps: Point the certificate to /system/etc/
The GPS HAL looks for the SSL certificates for use with the
navigation system. It looks at /data/system/gps/mrvlgps for the .pem file
but it won't be present so it shall fail silently.
Luckily, the CER_PATH option is present so it should be possible to point it
to the .pem file in the /system/etc directory.

Change-Id: I55275232048d4a36fe8f21aca7dd236d33a9909e
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>