gps: Enable power save mode
It needs further verification that this impacts the battery life when idle.

The situation is that when avdd18 regulator (the regulator to the GPS chip)
is disabled, the idle battery drain goes down to about 1% per hour with my
current battery. It's usually about 2-3% per hour.

The GPS HAL does not seem to disable the regulator but instead sends a
command to the chip to "turn off" based on the kernel logs.
Interestingly enough, there is a power save option in the config
which is currently set to off.

If this setting works, then it seems that expanding the wrapper
to disable the regulator on enable/disable wouldn't be needed.

Change-Id: Ic24fb4f8c28a22f09c608730eb1093af9b59b431
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>