kanas: Update libomx_avcdec_*_sprd.so
This is accompanied by the patch that switches over, or forces to use,
the sc8830 (or scx30g2 in Android N) version of the libstagefright module.

This doesn't increase the decoding performace but it seem to work much
better in handling cases when the VSP hardware is too slow to decode
a frame within the deadline.

Change-Id: I86bd7fbc32dd829680a2e37e7f75018d2ed93d1a
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>