kanas: Compatibility with setup-makefiles.sh and extract-files.sh; Remove unused cruft
I decided to support and to use the  setup-makefiles.sh and the extract-files.sh
scripts after I found a LineageOS wiki page that documents how to use it.

Also, we're removing some libraries that were not essential either because no one
depends on them or we could build it with the source code that we have.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>