kanas: patch libMali.so to use libion_sprd.so rather than libion.so
Starting Android 9, libion.so will probably be built along side the
libion_sprd library that some Spreadtrum modules used
(including this libMali.so). Fortunately most of them can be compiled/built
from source to directly use the libion_sprd.so

Unfortunately, libMali.so does not have a source.
Luckily, we have patchelf to do the manual patching to that library though
it may have increased the size of the library primaily due to padding 'X'
in the library.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>